refactor(meter-values): split CoherentMeterValuesGenerator into three modules (issue #1936 item i)
Split the 820 LOC `CoherentMeterValuesGenerator.ts` (introduced by PR
#1935) into three focused modules per issue #1936 item (i):

- `CoherentSampleComputer.ts` (311 LOC): physics chain V→P→I→ΔE→SoC with
  INV-1/2/3 by construction (guard bundle, ramp factor, voltage noise,
  EVSE/EV/capacity clamps, integer-rounded emission, energy accrual,
  monotone SoC saturation), plus `advanceEnergyRegister` which the
  builder invokes once per sample so `meterStop` stays correct even when
  `Energy.Active.Import.Register` is not in the configured MeterValues.
- `CoherentMeterValueBuilder.ts` (364 LOC): OCPP {@link MeterValue}
  assembly — phase family classifier, cross-measurand emit order
  (SoC→Voltage→Power→Current→Energy), within-measurand phase rank
  (no-phase → L1/L1-N → L2/L2-N → L3/L3-N → L1-L2 → L2-L3 → L3-L1 → N),
  kilo/unit divider, template resolution, and the `buildCoherentMeterValue`
  entry point that orchestrates `computeCoherentSample` +
  `advanceEnergyRegister` + per-template emission.
- `CoherentMeterValuesGenerator.ts` (204 LOC, shrunk from 820): session
  lifecycle (`createCoherentSession`, `CreateSessionOptions`), PRNG
  helpers (`createStreamPrng` with the `tx:`-namespaced deriveSeed leg,
  `resolveRootSeed`), the `isCoherentModeActive` type guard, and the
  module-scope WeakMap runtime state (`SessionRuntime`,
  `getSessionRuntime`, `disposeCoherentSessionRuntime`) accessed by the
  computer.

The three modules stack: Generator (WeakMap + PRNG helpers) ← Computer
(imports `getSessionRuntime`, `createStreamPrng`) ← Builder (imports
`computeCoherentSample`, `advanceEnergyRegister`, `ROUNDING_SCALE`,
`CoherentSample`, `ComputeSampleOptions`). No cycles.

The barrel (`meter-values/index.ts`) re-exports `buildCoherentMeterValue`
+ `BuildVersionedSampledValue` from the builder and
`createCoherentSession` / `isCoherentModeActive` / `resolveRootSeed`
from the generator so external consumers
(`CoherentMeterValuesManager`, `OCPPServiceUtils`, `OCPP16ServiceUtils`)
keep their existing imports.

Preserved invariants (per mission constraints):
- `__injectCoherentSession` NODE_ENV production-guard throw
  (unchanged — lives on `CoherentMeterValuesManager`).
- Session-snapshot reads for `voltageOutNominal` / `currentType` /
  `numberOfPhases` (still read from `session.*` inside
  `computeCoherentSample`).
- `tx:` PRNG namespace on the transactionId leg of `createStreamPrng`
  (preserves the XOR self-inverse guard).
- non-finite-`intervalMs` defensive zero-sample guard
  preserved verbatim in `computeCoherentSample`.

Test file imports updated to point at the new module boundaries; no test
logic changes. Test invariant `fail: 0, skipped: 6` holds.

Closes issue #1936 item (i).
